+++ apps/web/src/services/cost-explorer/components/BudgetDetailNotifications.vue
@@ -22,7 +22,7 @@ import BudgetDetailNotificationsChannelTable
 import BudgetDetailNotificationsConditionSettingModal
     from '@/services/cost-explorer/components/BudgetDetailNotificationsConditionSettingModal.vue';
 import { useBudgetDetailPageStore } from '@/services/cost-explorer/stores/budget-detail-page-store';
-import { PROJECT_ROUTE } from '@/services/project/routes/route-config';
+import { PROJECT_ROUTE } from '@/services/project/routes/route-constant';
 
 
 interface Props {
