+++ apps/web/src/services/notification/modules/NotificationChannelList.vue
@@ -111,7 +111,7 @@ import type { ChannelItem, EnrichedProtocolItem, ProtocolItem } from '@/services
 import { MY_PAGE_ROUTE } from '@/services/my-page/route-config';
 import NotificationChannelItem from '@/services/notification/modules/notification-channel-item/NotificationChannelItem.vue';
 import { PROTOCOL_TYPE } from '@/services/notification/modules/notification-channel-item/type';
-import { PROJECT_ROUTE } from '@/services/project/routes/route-config';
+import { PROJECT_ROUTE } from '@/services/project/routes/route-constant';
 
 export default {
     name: 'NotificationChannelList',
